+++ src/main/java/org/la4j/vector/Vectors.java
@@ -27,8 +27,11 @@ import java.math.BigDecimal;
 import java.math.RoundingMode;
 
 import org.la4j.LinearAlgebra;
+import org.la4j.factory.Factory;
 import org.la4j.io.MatrixMarketStream;
 import org.la4j.io.SymbolSeparatedStream;
+import org.la4j.vector.builder.TerminalVectorBuilder;
+import org.la4j.vector.builder.VectorBuilder;
 import org.la4j.vector.functor.VectorAccumulator;
 import org.la4j.vector.functor.VectorFunction;
 import org.la4j.vector.functor.VectorPredicate;
@@ -508,6 +511,7 @@ public final class Vectors {
      *
      * @return a MatrixMarket stream source
      */
+    @Deprecated
     public static VectorSource asMatrixMarketSource(InputStream in) {
         return new StreamVectorSource(new MatrixMarketStream(in));
     }
@@ -519,6 +523,7 @@ public final class Vectors {
      *
      * @return a symbol separated stream source
      */
+    @Deprecated
     public static VectorSource asSymbolSeparatedSource(InputStream in) {
         return new StreamVectorSource(new SymbolSeparatedStream(in));
     }
@@ -531,8 +536,20 @@ public final class Vectors {
      *
      * @return a symbol separated stream source
      */
+    @Deprecated
     public static VectorSource asSymbolSeparatedSource(InputStream in, String separator) {
         return new StreamVectorSource(new SymbolSeparatedStream(in, separator));
+    }
+
+    /**
+     * Creates a new vector builder instance of given {@code factory}.
+     *
+     * @param factory the builder's factory
+     *
+     * @return a factorised vector builder
+     */
+    public static VectorBuilder asBuilder(Factory factory) {
+        return new TerminalVectorBuilder(factory);
     }
 
     /**