+++ CHANGELOG.md
@@ -9,6 +9,7 @@ la4j-0.5.0 `May 2014`
  * New entity: `MatrixBuilder`
  * New matrix/vector method `mkString()` for customized to-string conversion
  * Bug fix in random generation of sparse matrices
+ * New entity: `VectorBuilder`
 
 la4j-0.4.9 `Jan 2014`
  * Bug fix in `align()` method for big sparse matrices (reported by Michael Kapper)